Improves CI workflow and adds Rust lints
Adds `-Dwarnings` to the `RUSTFLAGS` environment variable in the CI workflow to treat warnings as errors.
This ensures that the build fails if there are any warnings, promoting cleaner and more robust code.

Also updates the clippy check to run on all targets and features.

Adds `file` and `libxdo-dev` dependencies to the CI environment to support required functionalities.